Add achievement list
User achievements can now be viewed, although it's not styled.

This is a big change to navigation, as it includes adding bottom tab navigation.

Each navigation tab also has a navigation stack. Currently we have the Home stack and the Achievements stack.

Later on the Achievements tab will probably turn into a Profile tab.