Set up login and physical device testing
Login screen is now stylized (it looks slick af).

I also edited server constant so I can test from physical device and properly make calls to localhost - this was a problem
before because on iOS you can't make direct calls to APIs via localhost.

I still need to figure out what's up with promises not being resolved though. I thought it was an issue with
the emulator, but it's persistent on real device as well.